+++ Sibusten.Philomena.Client/Images/Downloaders/PhilomenaImageFileDownloader.cs
@@ -52,7 +52,7 @@ namespace Sibusten.Philomena.Client.Images.Downloaders
                     {
                         Current = streamProgress.BytesRead,
                         Total = streamProgress.BytesTotal,
-                        Action = downloadItem.Id.ToString(),
+                        Action = $"Downloading image {downloadItem.Id}",
                     });
                 });